Do not show token popup menu while dragging map
_Fix applies to both PointerTool and StampTool_

When releasing the mouse, we grab the current drag state before it gets cleared, so that we know whether the user was
dragging the map. This is in lieu of a more comprehensive change to `PointerTool` and `StampTool` that would isolate
different user actions to ensure this kind of thing can't happen.