Cache hard FoW, soft FoW and clear area, along with visible area
This avoids the need to `.intersect()` the visible area with the cleared area on every frame, which is relatively
expensive both in terms of speed and memory allocations. This is used in the `FogRenderer` to great effect, and to
lesser effect - due to the subsequent transformation - in `ZoneRenderer#showBlockedMoves()`.

Also add clarifying doc comment to ZoneView#getExposedArea() that it is only to be used when the Zone has fog enabled.