Update performance window
The font is now monospace so that the various lines are aligned with one another. The text area is also now uneditable
so it doesn't have to worry about things like tracking undo state.

The major change here is that text is added to the performance window via `JTextArea#append(String)` instead of
`JTextComponent#setText(String)`. The former does not require us to do string concatenation (which is expensive when the
contents are large), and means the text area's document does not have to reallocate a lot of its internal state.