Add ability to look up entries by index
The new `LookupTable#getEntryByIndex(int)` provides fast lookup of a table using an integer index.

This very directly helps in expressing the implementation of the `getTableEntryAtIndex()` macro function. But there is
no real performance difference here since the old and new implementations are effectively identical.

Where it really helps is with the `resetTablePicks()` macro function, which had to use the incredibly inefficient
`LookupTable#reset(String)` to reset entries. This method has also sorts of efficiency problems:
1. The signature bakes in inefficiency since each input `String` has to be parsed to a `int` (and in this case, a boxed
   `Integer`). So if a callers has an `int` ready to go, they first have to stringify it so that `reset(String)` can
   parse it back and work with it.
2. The parsed indices were not used to perform nigh-instantaneous lookups of the entry list. Instead, they were
   converted to a `Set` of boxed `Integer`. Then the entire entry list was iterated over and checked against the set.
   This is terrible news when the entry list gets long, as the check against a set is slower than an index lookup and
   because most entries will typically not be the target of the reset anyways.
3. The entry list is copied over into a new list while iterating. This is completely pointless as `LookupTable` is
   anything but thread safe, so couldn't possibly encounter sane concurrent modifications anyways.

Instead of all that cruft, `resetTablePicks()` converts its own input strings to `int` (not boxed!) which are then used
with `LookupTable#getEntryByIndex(int)` to get an entry reference. That entry is then reset directly.

The existing `LookupTable#getPickOnceLookup(String)` is now implemented on top of `LookupTable#getEntryByIndex(int)` for
consistency and clarity. It also does not redundantly write the entry back to `entryList` after setting the pick flag.

Also rename `LookupTable#reset()` to `LookupTable#resetPicks()` so that intent is clearer.