Improve standard table lookups and `getTableEntry()` especially
More baked in inefficiencies are gone. The old `LookupTable#getLookupDirect(String)` had these problems:
1. It had to run an expression parser to get an integer result to then lookup in tables. There is no way to skip this
   overhead and just supply an `int`.
2. It would constrain the result so it lies within the table's minimum and maximum values. Except... `LookupTable`
   doesn't actually keep track of that so it has to iterate over the entire list to look for the min and max values to
   clamp the roll result within.
3. A knock-on effect of (2) is that callers are forced to validate the result to make sure they didn't get a bogus
   result. Also interesting that gaps in the table are not afforded special treatment, returning `null` if encountered,
   so callers have to be prepared for that possiblity anyways. So this constraining operations doesn't help anyone.

And for `getTableEntry()`, which is essentially a wrapper around the above, point (3) comes to bite. In order to
validate the result, the macro implementation parses the roll as an `int`! So if we were to pass in a roll expression as
a string, the `LookupTable` would happily evaluate it and waste its time find a result, only for the macro
implementation to throw a `NumberFormatException` and lose the results!

All this is fixed by a new lookup method: `LookupTable#getEntryByRollResult(int)`. It does not expression evaluation,
leaving that responsibility to the caller. It also never returns an entry whose range does not encompass the input.

This method is also useful in `ImageSupport#getTokenImage(Token, ImageObserver[])` and `Token#getTokenImageAssetId()`.
Both these methods have ready access to an `int` - the token facing - but previously had to convert it to `String` so
that `LookupTable#getLookupDirect(String)` could evaluate it back to an `int` to do the look up. Use of the new method
also allows these methods to directly state that they do not want index-based lookups even if the table happens to be a
pick once table. Just imagine pick-once behaviour for image tables! For bonus points, neither method has to handle
`ParserException` anymores.

The existing `LookupTable#getStandardLookup(String)` is now implemented on top of `LookupTable#getEntryByRollResult(int)`
for consistency and clarity.