Split up LookupTable#getDefaultRoll() to streamline lookups
The pick-once branch of `#getDefaultRoll()` has been replaced with `#getRandomPickOnce()` which returns the index of a
random unpicked entry. It does not convert the index into a string, as `#getDefaultRoll()` had to, which means the index
can be directly used to lookup the entry rather than passing through the expression parser. Additionally, the randomness
for selecting the random index is taken from a plain `Random` instance rather than being prefixed with `"d"` and being
sent to the expression parser.

For the standard roll branch, `#getDefaultRoll()` has been replaced with `#getStandardDefaultRoll()`. This is
functionally the same as the original implementation, but is slightly streamlined by fast-tracking the empty case and
avoiding the null checks in the tight loop.

When calling `#getLookup(String)` with a `null` roll, we no longer produce a default roll expression before running the
pick once or standard lookup logic. Instead, we allow `#getPickOnceLookup(String)` and `#getStandardLookup(String)` to
handle `null` rolls, which they can do much more efficiently (see above for notes on `#getRandomPickOnce()`.

With these changes, `#getDefaultRoll()` itself is no longer needed as `#getRoll()` is the only place that calls it and
will simply return the result. So `#getDefaultRoll()` has been merged into `#getRoll()`.