Refactor the boolean token overlay hierarchy
Classes in the hierarchy are either abstract or final, so there is no deep nesting anymore. The only real inheritance
that is used now is when several types share a common structure, so can be handled in identical ways during rendering,
persisting, etc. This makes the leaf classes generally quite simple, and in general it's easier to follow the rendering
process for each implementation.

The DTOs for overlays have been reworked to match the new inheritance tree. `BooleanTokenOverlayDto` no longer supplies
all possible fields for all overlay implementations, but supplies only the common fields and a union of messages, one
for each unique overlay structure.

This is the new hierarchy:
- `BooleanTokenOverlay`
  - `AbstractShapeTokenOverlay`
    - `CrossTokenOverlay`
    - `DiamondTokenOverlay`
    - `OTokenOverlay`
    - `TriangleTokenOverlay`
    - `XTokenOverlay`
    - `YieldTokenOverlay`
  - `AbstractFlowShapeTokenOverlay`
    - `FlowColorDotTokenOverlay`
    - `FlowColorSquareTokenOverlay`
    - `FlowDiamondTokenOverlay`
    - `FlowTriangleTokenOverlay`
    - `FlowYieldTokenOverlay`
  - `ColorDotTokenOverlay`
  - `CornerImageTokenOverlay`
  - `FlowImageTokenOverlay`
  - `ImageTokenOverlay`
  - `ShadedTokenOverlay`

Compare with the old hierarchy:
- `BooleanTokenOverlay`
  - `ImageTokenOverlay`
    - `CornerImageTokenOverlay`
    - `FlowImageTokenOverlay`
  - `ShadedTokenOverlay`
  - `XTokenOverlay`
    - `ColorDotTokenOverlay`
    - `CrossTokenOverlay`
    - `DiamondTokenOverlay`
    - `FlowColorDotTokenOverlay`
      - `FlowColorSquareTokenOverlay`
      - `FlowDiamondTokenOverlay`
      - `FlowTriangleTokenOverlay`
      - `FlowYieldTokenOverlay`
    - `OTokenOverlay`
    - `TriangleTokenOverlay`
    - `YieldTokenOverlay`