Encapsulate Pen using readResolve() and writeReplace()
`Pen` now has a simple invariant: the foreground mode is transparent if and only if the foreground paint is null, and
the background mode is transparent if and only if the background paint is null. `Pen#readResolve()` guarantees this
invariant holds when loading legacy pens, and `Pen#writeReplace()` guarantees that serialized pens will have the same
properties and can be properly handled by a subsequent `Pen#readResolve()`. Callers are also no longer able to check or
modify the foreground and background modes, instead only setting the desired paints.

When loading legacy pens, a non-positive opacity is replaced with `1`. Callers thus no longer need to check for an
opacity of zero.

Legacy colors are no longer publicly accessible. Callers can just use the modern paints.