Avoid object state for passing data around HaloRenderer
For `haloPartMap` and `haloFacingAngle`, this is a simple matter of just not setting the fields and instead passing the
values as parameters to `renderHaloPart()` and elsewhere. This actually exposes a potential NPE weakness in the current
implementation in case there is no parent `Halo`, as is the case for legacy halos. This will be tightened up later.

For `haloScaleFactorMap`, we actually need to remember the scale factor in one place to use it later during the reverse
halo loop. For this reason, we introduced the `RenderablePart` record that contains all information needed to render a
specific `HaloPart`, including the scale factor and other fields.

For `haloLineWidthPreference`, we combine this with the halo's own line width quite early, and also include it in
`RenderablePart` so that the effective line width calculation does not need to be duplicated.

We also lifted the calculation of the halo shape type up into `RenderablePart` so we can calculate it once early on and
then reuse the results. This includes using the default shape when the part's shape is `null`, and converting `TOKEN` to
either `CIRCLE` or `GRID`. Similarly, we lift the calculation of the facing angle out of the reverse loop and make it
part of `RenderablePart`.

Only the `haloClipArcByCircleShape` field remains in this regard, but avoiding it requires further refactoring.