Rework how halo shapes and masks are computed
The core of this commit is splitting the calculation of the halo shape in `getHaloGeometricShape()` from the calculation
of the special mask used for arcs. To avoid excessive duplication, this required removing `transformHaloShape()` and
replacing it with a more generic `AffineTransform`-returning `buildTransform()` that takes similar parameters but does
not itself transform a shape. Callers can easily apply this transform with no increase in cognitive complexity.

During rendering, the mask is now injected as a method parameter to `paintHaloPart()`, freeing that method from deciding
what to use as a mask.

With these changes, we no longer need the `haloClipArcByCircleShape` field, avoiding any possible issues with state
bleeding over from one part to the next.

We have also lifted the calculation of the unit shapes up into `renderHaloPart()`. This makes it clear that the unit
shape is only useful for geometric shape types (`CIRCLE`, `ARC`, `PIE`, `CHORD`, `TRIANGLE`, `SQUARE`, `POLYGON`, and
`STAR`) and the same unit shape is used for mini shapes or regular geometric shapes.