Add pick-once functionality to the Edit Table dialog
The new "Pick Once Table" checkbox allows the user to set a table as pick once without resorting to macros. When a table
is pick once:
- The roll option is disabled as it is not used.
- A new "Picked?" columns is shown in the table to indicate whether a raw has already been picked. The user can manually
  select or unselect these to mark specific entries as picked.
- The new "Reset Picks" button will be enabled, allowing the user to clear all the "Picked?" checkboxes.

This also means that editing a table will no longer clear the picked entries unless the user manually clears them or
changes the table to not be a pick once table.