Fix AddResourceDialog double-install
`GenericDialog` automatically commits the dialog since the content component is an `AbeillePanel` with a bound model.
This means we don't have to manually wire up the INSTALL button to the `#commit()` method, and in fact doing so is what
causes resource libraries to be installed doubly.

Also removes the redundant `AddResourceDialog#model` field - as an `AbeillePanel`, the `getModel()` method already
returns this same value.