Make AbeillePanel#bind() more robust by not throwing exceptions
This isn't the right situation to throw an exception, since it is at least as likely to indicate a failure for previous
code to unbind the model rather than being a problem of the current caller double-binding. Throwing an exception here
has the potential to crash MT and even the OS as the exception is reported in error dialogs on repeat.

This case still indicates a logical error, though, so log it as such. But after that, just unbind the old model and
carry on.