Remove RessourceManager#main()
This tool detected missing icons in the `Rod Takehara` icon theme. But this can be better done by not using a
`HashMap<>` to map `Icons` to paths, but instead using `switch()` expressions that are lighterweight and wouuld enforce
exhaustiveness, making it explicit in the source code which icons are not supported.