GenericDialog now blocks dialog close when the commit() method fails
This case only applies for AFFIMRATIVE actions such as pressing a dialog's OK button. By cancelling the dialog close, we
give the user a chance to correct and validation errors.

Fixes the **Add Resource** and **Edit Token** dialogs that have both been changed to better leverage the GenericDialog,
but as a result suffered from closing upon validation errors.