Move library downloads out of AddResourceDialog into LibraryUtils
`LibraryUtils` handles the downloading and parsing of the library list, representing the results as a collection of
`Library` records. When parsing the library list, it is much more strict about required formats, not permitted empty
names, paths, or sizes, and not allowing paths that result in invalid URLs.

`LibraryUtils` is also responsible for downloading and installing such a collectionm, though as a separate operation so
that the caller can filter the results if desired.

With the change in responsibilities, the test for `DownloadListWorker` in `AddResourceDialogTest` have been moved to the
new `LibraryUtilsTest` and test `LibraryUtil#parseLibraryList(String)`. The other tests in `AddResourceDialogTest`
tested that `LibraryRow` initialized itself properly, but that is no longer a real concern now that it just wraps a
`Library` and `Library` is basic record.

Error reporting works differently now. Previously, we would show error dialogs to the user any time anything went wrong,
often regarding things that have nothing to with them or the current MT installation. E.g., we would show a popup if the
server returned an invalid path for a library. What we do now is to collect all the failed installations - regardless of
why they failed - and show the user one dialog to inform them which of their desired libraries could not be
installed. As a result, some i18n keys have been removed.