Show the author for RPTools libraries in Add Resource to Library
The library author is displayed in a new column of the RPTools table. This not only distinguishes it, but allows the
user to sort by author in addition to library name or size if they so desire.

For backwards compatibility, the author in the download library list is optional (though in practice it already exists
in the listing). It's also just not a critical detail, so it's okay if it is missing. If there is no author, the dialog
will display the name as "unknown" (`dialog.addresource.unknownAuthor`).

With three equal-width columns, there is no longer enough room to display most library names. So we instead assign 60%
of the table width to the library name, 25% to the author name, and 15% to the size. The user can resize these if
needed, but in practice this appears to provide sufficient room. Once the user resizes any columns, the proportions they
set will be preserve even if the table is subsequently resized.