Replace JList with JTable for AddResourceDialog RPTools tab
With the `JTable`, we now have separate columns for the library name and the size. The user can sort by either column by
clicking on the column header.

With this change, there is no more need for `LibraryRow` as we don't depend on the `#toString()` method of the rows as
was needed with `JList`.