Add stronger typing for AddResourceDialog UI
The raw `JList` is now a `JList<LibraryRow>`, ensuring that it never contains a `String.` This completely avoids
possibilities such as seen in #343.

This is accomplished by moving the informational and error messages out of the `JList` and into a dedicated
`JLabel`. This label is only displayed when there is a message to show the user and is completely hidden otherwise.