Override FlatLaf and show table borders by default
It is now update to individual tables to decide whether they need borders. Most table leave the borders disabled. The
only tables to enable them are the ones that need to support editing by the user, including the addition of new rows:
- The Speech table in the **Edit Token** dialog.
- The table in the **Edit Table** dialog.

While there are other editable tables, they are more specialized and so do not have the same usability issues.