Mark editable tables as not allowing row selection
This makes is much clearer which cell is being edited as the UI for indicating that is not obscured by the row
highlight. Since uneditable tables don't need such feedback, and since row selection is sometimes a key feature for such
tables, the highlighting has been left in place for those tables. The only affected tables are thus the **Edit Token**
dialog's speech table, and the **Edit Table* dialog's table.