Add utility for customizing JTable header cells.
`JTable` header cells are difficult to customize without messing up the theme since the default (and themable)
implementation for header cells is not publicly accessible and cannot be instantiated. Using `DefaultTableCellRenderer`
would cause the header cell to be themed as a regular table cell, not a header cell.

The new `TableCellRendererDecorator` can wrap any `TableCellRenderer` (in particular, the header cell renderer from an
existing table) and will modify the resultsin of the `#getTableCellRendererComponent()` to achieve the desired
customization.

Only vertical and horizontal text alignment are current supported, and requires the assumption that the decorated
renderer produces `JLabel` instance (which is true in practice).