Move the two main rows into a split pane
This allows the user to adjust the spacing if they want to focus more on property types or properties proper. The
default split is 30/70, giving ever so slightly more vertical space to the property table.

To get the propertions to make sense, this required removing some of the hardcoded preferred sizes. Those were not
actually needed anymore, so it's good for them to be gone anyways. An extra VSpacer was required to get the properties
table to stretch but not its button panel.