Clean up various board-related logic
In `Zone`, put all the board getters & setters together. Rename `setBoard(Point)` to `setBoardPosition(Point)` to avoid
confusion with the more general `setBoard(MD5Key, Point)`. Modify `setBoard(MD5Key, Point)` to be the one-true-place to
update board settings, with this method firing a `BoardChanged` event and the other board-related setters delegating to
it.

Include the x/y scale in with the other board parameters. Include it in the `BoardChanged` event and `SetBoardMsg` DTO.

Deprecate `Zone#boardChanged` in favour of the renderer keeping track of this information. Remove the corresponding
`boardChanged` from `ZoneDto`. Since this is a serialized field, it has not been removed.