Pull ExportDialog responsibility out of Campaign
The settings are still in `Campaign`, but the dialog itself is created on-demand by the corresponding actions. This
defers dialog creation to after app initialization, meaning it is now themes as the rest of the application.

The two actions have been rolled into a single implementation as they both have to handle the "save as" case.

With `ExportDialog` now being a singleton, there is no more need for the static state. The static `AbeillePanel` remains
as a kludge.