Reimplement InitiativeListCellRenderer
It is now based on `GridLayoutManager` instead of `MigLayout` so that we have more control over how the individual
components are laid out. The most important aspect is that we no longer need to depend on HTML to get the initiative on
its own line, which in turn allows the name text to overflow using ellipsis.

The token name and initiative are now in separate labels instead of being in a single label. These two labels are packed
into the new `TextPanel` component. There are two variants of `TextPanel`: one that puts both labels in one row, with
the initiative aligned right; the other that puts both labels in column, with the initiative aligned left.

The token image is no longer set as a custom icon drawn for a `JLabel`, but is rendered to a new `TokenImagePanel`
component that is just a `JPanel` with custom painting behaviour. With this, there is no more need to produce a scaled
`BufferedImage` as the token image can be scaled in realtime without trouble. This also means we have no need to cache
the scaled image in the `TokenInitiative`, eliminating a bug where the token image is not updated in the initiative list
after editing the token.

The `TokenImagePanel` and `TextPanel` are packed into another panel that forms the "label panel" part of the initiatve
list cell component. The flat label background is drawn based on this component's bounds. When a token is holding, the
label panel is aligned right, otherwise it is aligned left. Additionally, the token image is placed in the third column
when holding, otherwise it is placed in the first column, so that the ordering of name vs token image is the same as
before.

The cell as a whole is now a two-column `JPanel`. The first column holds the current initiative icon, while the second
contains the label panel described above.